mem: split mem.h
