+++ vita3k/mem/src/mem.cpp
@@ -15,7 +15,9 @@
 // with this program; if not, write to the Free Software Foundation, Inc.,
 // 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-#include <mem/mem.h>
+#include <mem/functions.h>
+#include <mem/state.h>
+
 #include <util/align.h>
 #include <util/log.h>
