+++ vita3k/net/include/net/types.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <mem/mem.h>
 #include <mem/ptr.h>
 
 enum SceNetProtocol {
